requirements.txt
numpy==1.24.3
faiss-cpu==1.8.0
google-genai==1.23.0
scikit-learn==1.3.0
packaging>=21.0 